refactored for context exercise
