+++ src/components/Favorites.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+
+const Favorites = () => {
+  return (
+    <div>
+        <h3>Hi from Favorites.js</h3>
+    </div>
+  );
+};
+
+export default Favorites;
