+++ src/components/FoxCard.js
@@ -1,5 +1,4 @@
 import React from "react";
-import "../App.scss";
 
 const FoxCard = ({ image }) => {
   return (
